WIP/RFC: Implement mipmaps
This implements mipmapping by adding a third argument to the
'INTERPOLATION' glir command.

It's not clear what the right way to support mipmapping is since setting
the number of mipmap levels is only supported in >= OpenGL ES 3.0 and
>= OpenGL 3.0. This discrepancy is handled by checking whether
'glTextureStorage2D' is available in the gl package.

Any comments are welcome.